+++ packages/components/src/components/IntlProvider/IntlProvider.ts
@@ -1,0 +1,2 @@
+import { I18nProvider } from "react-aria-components";
+export const IntlProvider = I18nProvider;